bpf: move metrics to new cilium_metrics_v2 map
With commit 38ab8f0f3 the format used to encode BPF metrics in the
cilium_metrics map has changed.

Specifically, the set of constants used to encode the direction of a
given metric has changed from:

  #define METRIC_INGRESS  1
  #define METRIC_EGRESS   2

to:

  #define METRIC_EGRESS   0
  #define METRIC_INGRESS  1
  #define METRIC_SERVICE  2

Since after an upgrade the cilium_metrics map is not deleted, upgrading
from a version which uses the old directions encoding to a version which
uses the new one causes "cilium bpf metrics list" to report incorrect
metrics, as all the keys already present in the map will be interpreted
following the new format.

This commit fixes this by introducing a new map for metrics,
cilium_metrics_v2, while also keeping the old one (cilium_metrics).

This allows to mitigate the upgrade impact by copying over the eventual
items of the old map, after they have been rewritten to match the new
format, into the new map.

This also partially mitigates the downgrade impact, as downgrading to a
version which does not use the new direction scheme will not cause
"cilium bpf metrics list" to report invalid metrics, although all
metrics collected with the new format in cilium_metrics_v2 will be lost.

Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>